% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{seedwt.multi.subsample}
\alias{seedwt.multi.subsample}
\title{A R dataset}
\format{An object of class \code{data.frame} with 2916 rows and 10 columns.}
\usage{
seedwt.multi.subsample
}
\description{
A R dataset
}
\details{
(details here; DIU)
}
\references{
Wang, Y., Stapleton, A. E., & Chen, C. (2018). Two-sample nonparametric stochastic order inference with an application in plant physiology. Journal of Statistical Computation and Simulation, 88(14), 2668-2683.
}
\keyword{datasets}
