% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{simu.ustat.pattern}
\alias{simu.ustat.pattern}
\title{simu.ustat.pattern function}
\usage{
simu.ustat.pattern(mean.prob.vec, effn.subs, n.rep = 10^2)
}
\arguments{
\item{mean.prob.vec}{a vector of length 2. Its first element represents the probability that a random observation from one subsample is less than the the one from another subsample..}

\item{effn.subs}{a vector contains two subsample sizes.}

\item{n.rep}{the total number of repetition.}
}
\value{
simu.tab a list of length n.rep. Each element of the list is a 2 by 2 matrix, showing the comparison results from function multi.freq.
}
\description{
This function create two independent subsamples of various subsample sizes, with a given probability vector.
}
\details{
each subsample is generated from a normal distribution, with an average generated from the mean.prob.vec.
}
\examples{
simu.ustat.pattern(c(0.8,0.2),c(5,8),n.rep=100)
}
\references{
Wang, Y., Stapleton, A. E., & Chen, C. (2018). Two-sample nonparametric stochastic order inference with an application in plant physiology. Journal of Statistical Computation and Simulation, 88(14), 2668-2683.
}
